HV:misc:fix "signed/unsigned conversion with cast"
Signed/unsigned conversion should add cast explicitily
or change the type of them to the same.

V1->V2:Fixed the 0U to 0UL because of the mistakes.
V2->V3:remove unsed macro

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/include/arch/x86/assign.h
@@ -9,7 +9,7 @@
 
 #include <ptdev.h>
 
-void ptdev_intx_ack(struct vm *vm, int virt_pin,
+void ptdev_intx_ack(struct vm *vm, uint8_t virt_pin,
 		enum ptdev_vpin_source vpin_src);
 int ptdev_msix_remap(struct vm *vm, uint16_t virt_bdf,
 		struct ptdev_msi_info *info);
